Refactored all event date/time functions
Updated addToCart function to include eventDate and eventTime keys. Enabled menu item links on page load if cart is not empty.  Instead of using hidden fields to store the last valid date/time entry, it will now be stored in the cart key in session storage. If a cart is present on page load, the event date/time input will be populating using the date/time that’s in the cart. Also, when the date/time inputs are modified with valid data, cart is updated with these values.