Removed function that add listeners to nav links
Also reverted changes to function that toggled the main nav panel opened or closed. I discovered that the panel is open is I press forward or back and return to the current page if I never clicked a link in the nav panel. So this issue needs to be resolved on the window level and not an a particular element.